python/pyproject.toml
# MIT License
#
# Copyright (c) 2025 Meher V.R. Malladi
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.

[build-system]
requires = ["scikit-build-core", "pybind11"]
build-backend = "scikit_build_core.build"

[project]
name = "rko_lio"
version = "0.0.1"
description = "Python bindings for rko lio"
dependencies = ["numpy", "typer", "pyyaml", "scipy", "tqdm"]
authors = [{ name = "Meher Malladi", email = "rm.meher97@gmail.com" }]

[project.optional-dependencies]
all = [
    "open3d",
    "rerun-sdk",
    "rosbags"
]

[project.scripts]
rko_lio = "rko_lio.cli:app"

[tool.scikit-build]
wheel.packages = ["rko_lio"]
build-dir = "build/{wheel_tag}"
cmake.source-dir = "pybind"
cmake.version = ">=3.28"
cmake.build-type = "Release"
build.verbose = true
editable.mode = "redirect"
editable.rebuild = true
editable.verbose = true

